# extracted/Asset-Manager/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSquad API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  clinicName: string;
  ownerName: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  secretary: "secretary",
  superadmin: "superadmin",
} as const;

export interface User {
  id: string;
  email: string;
  role: UserRole;
  clinicId: string;
  name: string;
  isBlocked: boolean;
}

export type ClinicStatus = (typeof ClinicStatus)[keyof typeof ClinicStatus];

export const ClinicStatus = {
  pending: "pending",
  active: "active",
  blocked: "blocked",
  deleted: "deleted",
} as const;

export type ClinicSubscriptionStatus =
  (typeof ClinicSubscriptionStatus)[keyof typeof ClinicSubscriptionStatus];

export const ClinicSubscriptionStatus = {
  trial: "trial",
  basic: "basic",
  premium: "premium",
  expired: "expired",
} as const;

export interface Clinic {
  id: string;
  name: string;
  ownerId: string;
  status: ClinicStatus;
  subscriptionStatus: ClinicSubscriptionStatus;
  trialEndDate: string;
  subscriptionPlan?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  clinic: Clinic;
  token: string;
}

export type SubscriptionPlanType =
  (typeof SubscriptionPlanType)[keyof typeof SubscriptionPlanType];

export const SubscriptionPlanType = {
  basic: "basic",
  premium: "premium",
} as const;

export type SubscriptionPaymentStatus =
  (typeof SubscriptionPaymentStatus)[keyof typeof SubscriptionPaymentStatus];

export const SubscriptionPaymentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  failed: "failed",
} as const;

export interface Subscription {
  id: string;
  clinicId: string;
  planType: SubscriptionPlanType;
  startDate: string;
  endDate: string;
  paymentStatus: SubscriptionPaymentStatus;
  amount: number;
  createdAt: string;
}

export type CreateSubscriptionBodyPlanType =
  (typeof CreateSubscriptionBodyPlanType)[keyof typeof CreateSubscriptionBodyPlanType];

export const CreateSubscriptionBodyPlanType = {
  basic: "basic",
  premium: "premium",
} as const;

export interface CreateSubscriptionBody {
  planType: CreateSubscriptionBodyPlanType;
  /** Base64 or URL of payment proof image */
  paymentProof?: string;
}

export type PatientGender = (typeof PatientGender)[keyof typeof PatientGender];

export const PatientGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface Patient {
  id: string;
  clinicId: string;
  name: string;
  phone: string;
  dateOfBirth?: string;
  gender: PatientGender;
  bloodType?: string | null;
  allergies?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreatePatientBodyGender =
  (typeof CreatePatientBodyGender)[keyof typeof CreatePatientBodyGender];

export const CreatePatientBodyGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface CreatePatientBody {
  name: string;
  phone: string;
  dateOfBirth?: string;
  gender: CreatePatientBodyGender;
  bloodType?: string | null;
  allergies?: string | null;
  notes?: string | null;
}

export interface PatientList {
  data: Patient[];
  total: number;
  page: number;
  limit: number;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export interface Appointment {
  id: string;
  clinicId: string;
  patientId: string;
  patientName: string;
  date: string;
  time: string;
  status: AppointmentStatus;
  type: string;
  notes?: string | null;
  fee?: number | null;
  createdAt: string;
}

export interface CreateAppointmentBody {
  patientId: string;
  date: string;
  time: string;
  type: string;
  notes?: string | null;
  fee?: number | null;
}

export type UpdateAppointmentBodyStatus =
  (typeof UpdateAppointmentBodyStatus)[keyof typeof UpdateAppointmentBodyStatus];

export const UpdateAppointmentBodyStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export interface UpdateAppointmentBody {
  status?: UpdateAppointmentBodyStatus;
  date?: string;
  time?: string;
  notes?: string | null;
  fee?: number | null;
}

export interface AppointmentList {
  data: Appointment[];
  total: number;
  page: number;
  limit: number;
}

export type FinanceType = (typeof FinanceType)[keyof typeof FinanceType];

export const FinanceType = {
  income: "income",
  expense: "expense",
} as const;

export interface Finance {
  id: string;
  clinicId: string;
  type: FinanceType;
  category: string;
  amount: number;
  description: string;
  date: string;
  createdAt: string;
}

export type CreateFinanceBodyType =
  (typeof CreateFinanceBodyType)[keyof typeof CreateFinanceBodyType];

export const CreateFinanceBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateFinanceBody {
  type: CreateFinanceBodyType;
  category: string;
  amount: number;
  description: string;
  date: string;
}

export interface FinanceList {
  data: Finance[];
  total: number;
  totalIncome: number;
  totalExpense: number;
}

export type FinanceSummaryMonthlyBreakdownItem = {
  month: number;
  income: number;
  expense: number;
};

export interface FinanceSummary {
  totalIncome: number;
  totalExpense: number;
  netProfit: number;
  monthlyBreakdown: FinanceSummaryMonthlyBreakdownItem[];
}

export interface DashboardSummary {
  totalPatients: number;
  newPatientsThisMonth: number;
  todayAppointments: number;
  completedAppointments: number;
  monthlyRevenue: number;
  monthlyExpenses: number;
  upcomingAppointments: number;
  subscriptionStatus: string;
  trialDaysLeft?: number | null;
}

export type ListPatientsParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type ListAppointmentsParams = {
  date?: string;
  status?: ListAppointmentsStatus;
  page?: number;
  limit?: number;
};

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export type ListFinancesParams = {
  type?: ListFinancesType;
  month?: number;
  year?: number;
};

export type ListFinancesType =
  (typeof ListFinancesType)[keyof typeof ListFinancesType];

export const ListFinancesType = {
  income: "income",
  expense: "expense",
} as const;

export type GetFinanceSummaryParams = {
  year?: number;
};