# extracted/Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClinicSquad API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new clinic admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /clinics/{clinicId}/appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            type: string
            enum: [scheduled, completed, cancelled, no_show]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /clinics/{clinicId}/appointments/{appointmentId}:
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/clinics/{clinicId}/activate:
    post:
      operationId: adminActivateClinic
      tags: [clinics]
      summary: SuperAdmin - activate clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /admin/clinics/{clinicId}/block:
    post:
      operationId: adminBlockClinic
      tags: [clinics]
      summary: SuperAdmin - block clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /admin/subscriptions/{clinicId}/confirm:
    post:
      operationId: adminConfirmSubscription
      tags: [subscriptions]
      summary: SuperAdmin - confirm payment and activate subscription
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        clinicName:
          type: string
        ownerName:
          type: string
      required: [email, password, clinicName, ownerName]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        clinic:
          $ref: "#/components/schemas/Clinic"
        token:
          type: string
      required: [user, clinic, token]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, secretary, superadmin]
        clinicId:
          type: string
        name:
          type: string
        isBlocked:
          type: boolean
      required: [id, email, role, clinicId, name, isBlocked]
    Clinic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        status:
          type: string
          enum: [pending, active, blocked, deleted]
        subscriptionStatus:
          type: string
          enum: [trial, basic, premium, expired]
        trialEndDate:
          type: string
          format: date-time
        subscriptionPlan:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, ownerId, status, subscriptionStatus, trialEndDate, createdAt]
    Subscription:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        planType:
          type: string
          enum: [basic, premium]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        paymentStatus:
          type: string
          enum: [pending, confirmed, failed]
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, clinicId, planType, startDate, endDate, paymentStatus, amount, createdAt]
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        type:
          type: string
        notes:
          type: string
          nullable: true
        fee:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, clinicId, patientId, patientName, date, time, status, type, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        type:
          type: string
        notes:
          type: string
          nullable: true
        fee:
          type: number
          nullable: true
      required: [patientId, date, time, type]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
          nullable: true
        fee:
          type: number
          nullable: true
    AppointmentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...